% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.pstp.R
\name{plot.pstp}
\alias{plot.pstp}
\title{Plot for an object of class "pstp"}
\usage{
\method{plot}{pstp}(x = object, state_ini = 0, ...)
}
\arguments{
\item{x}{A fitted pstp object as produced by presmTP.}

\item{state_ini}{Initial state of the transition. Defaults to state_ini=0}

\item{...}{For future methods.}
}
\value{
No value is returned.
}
\description{
It draws the estimated probabilities.
}
\examples{
res<- presmTP(data = colonIDM, s = 365,method = "uns")
plot(res)

}
\author{
Gustavo Soutinho, Luis Meira-Machado, Pedro Oliveira.
}
